Create student.java
This is a Java class named "student" that defines methods for performing basic CRUD (Create, Read, Update, Delete) operations on a MySQL database table named "people".

The class has the following methods:

get_student_details(): This method prompts the user to enter their name and ID and stores the values in the name and id instance variables respectively.

insert_details(): This method creates a new record in the "people" table with the name and id values obtained from the get_student_details() method.

update_student_name(): This method prompts the user to enter their ID and their corrected name, and updates the corresponding record in the "people" table with the new name.

delete_student_record(): This method prompts the user to enter their ID and deletes the corresponding record from the "people" table.

search_details(): This method prompts the user to enter their ID and retrieves and prints the corresponding record from the "people" table.

The class uses the JDBC API to connect to the MySQL database and execute SQL statements. It creates an instance of the dbconnection class to establish the connection with the database and calls the closeConnection() method to close the connection after executing the SQL statement. The PreparedStatement class is used to execute parameterized SQL statements to prevent SQL injection attacks. The Scanner class is used to read input from the user.